compile again

+++ db/queryoptimizer.h
@@ -210,7 +210,7 @@ namespace mongo {
     template<class T>
     class our_priority_queue : boost::noncopyable { 
         vector<T> v;
-        int t;
+        size_t t;
     public:
         our_priority_queue() { 
             v.reserve(4);
